refactor

+++ Assets/_Scripts/HealthBar.cs
@@ -10,14 +10,14 @@ public class HealthBar : MonoBehaviour
     public Image fill;
     [SerializeField] HeroKnight player;
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
         fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
